+++ src/integrations/index.ts
@@ -1,1 +1,2 @@
 export * from './workItems'
+export * from './jira'